--- public/sitemap.xml
@@ -1,83 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-  <url>
-    <loc>https://mermer.com.tw</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/hiring</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/knowledge-management</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/hiring#job001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/hiring#job002</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/hiring#job003</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/hiring#job004</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-    <url>
-    <loc>https://mermer.com.tw/hiring#job005</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-    <url>
-    <loc>https://mermer.com.tw/hiring#job006</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-    <url>
-    <loc>https://mermer.com.tw/hiring#job007</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/en/knowledge-management/km-20230719001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/en/knowledge-management/km-20230721001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/en/knowledge-management/km-20230808001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/en/knowledge-management/km-20230809001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-    <url>
-    <loc>https://mermer.com.tw/en/knowledge-management/km-20230810001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-      <url>
-    <loc>https://mermer.com.tw/en/knowledge-management/km-20230818001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-   <url>
-    <loc>https://mermer.com.tw/en/knowledge-management/km-20230904001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-   <url>
-    <loc>https://mermer.com.tw/en/knowledge-management/km-20230821001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/en/knowledge-management/km-20230823001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-  <url>
-    <loc>https://mermer.com.tw/en/knowledge-management/km-20230830001</loc>
-    <lastmod>2023-09-22</lastmod>
-  </url>
-</urlset>
