+++ src/constants/config.ts
@@ -98,232 +98,6 @@ export const partnersContent = [
   },
 ];
 
-export const jobList = [
-  {
-    anchor: 'job001',
-    jobTitle: 'HIRING_PAGE.JOB_NAME_1',
-    details: [
-      'HIRING_PAGE.JOB_DETAIL_FULL_TIME',
-      'HIRING_PAGE.JOB_DETAIL_SALARY_1',
-      'HIRING_PAGE.JOB_DETAIL_LOCATION_1',
-    ],
-    descriptions: [
-      'HIRING_PAGE.JOB_DESCRIPTION_1',
-      'HIRING_PAGE.JOB_DESCRIPTION_2',
-      'HIRING_PAGE.JOB_DESCRIPTION_3',
-      'HIRING_PAGE.JOB_DESCRIPTION_4',
-      'HIRING_PAGE.JOB_DESCRIPTION_5',
-      'HIRING_PAGE.JOB_DESCRIPTION_6',
-      'HIRING_PAGE.JOB_DESCRIPTION_7',
-      'HIRING_PAGE.JOB_DESCRIPTION_8',
-    ],
-    requirements: [
-      'HIRING_PAGE.JOB_REQUIREMENT_1',
-      'HIRING_PAGE.JOB_REQUIREMENT_2',
-      'HIRING_PAGE.JOB_REQUIREMENT_3',
-      'HIRING_PAGE.JOB_REQUIREMENT_4',
-      'HIRING_PAGE.JOB_REQUIREMENT_5',
-      'HIRING_PAGE.JOB_REQUIREMENT_6',
-      'HIRING_PAGE.JOB_REQUIREMENT_7',
-    ],
-  },
-  {
-    anchor: 'job002',
-    jobTitle: 'HIRING_PAGE.JOB_NAME_2',
-    details: [
-      'HIRING_PAGE.JOB_DETAIL_FULL_TIME',
-      'HIRING_PAGE.JOB_DETAIL_SALARY_1',
-      'HIRING_PAGE.JOB_DETAIL_LOCATION_1',
-    ],
-    descriptions: [
-      'HIRING_PAGE.JOB_DESCRIPTION_1',
-      'HIRING_PAGE.JOB_DESCRIPTION_2',
-      'HIRING_PAGE.JOB_DESCRIPTION_3',
-      'HIRING_PAGE.JOB_DESCRIPTION_4',
-      'HIRING_PAGE.JOB_DESCRIPTION_5',
-      'HIRING_PAGE.JOB_DESCRIPTION_6',
-      'HIRING_PAGE.JOB_DESCRIPTION_7',
-      'HIRING_PAGE.JOB_DESCRIPTION_8',
-    ],
-    requirements: [
-      'HIRING_PAGE.JOB_REQUIREMENT_8',
-      'HIRING_PAGE.JOB_REQUIREMENT_9',
-      'HIRING_PAGE.JOB_REQUIREMENT_10',
-      'HIRING_PAGE.JOB_REQUIREMENT_11',
-      'HIRING_PAGE.JOB_REQUIREMENT_4',
-      'HIRING_PAGE.JOB_REQUIREMENT_5',
-      'HIRING_PAGE.JOB_REQUIREMENT_6',
-      'HIRING_PAGE.JOB_REQUIREMENT_7',
-    ],
-  },
-  {
-    anchor: 'job003',
-    jobTitle: 'HIRING_PAGE.JOB_NAME_3',
-    details: [
-      'HIRING_PAGE.JOB_DETAIL_FULL_TIME',
-      'HIRING_PAGE.JOB_DETAIL_SALARY_2',
-      'HIRING_PAGE.JOB_DETAIL_LOCATION_1',
-    ],
-    descriptions: [
-      'HIRING_PAGE.JOB_DESCRIPTION_9',
-      'HIRING_PAGE.JOB_DESCRIPTION_27',
-      'HIRING_PAGE.JOB_DESCRIPTION_28',
-      'HIRING_PAGE.JOB_DESCRIPTION_29',
-      'HIRING_PAGE.JOB_DESCRIPTION_30',
-      'HIRING_PAGE.JOB_DESCRIPTION_31',
-      'HIRING_PAGE.JOB_DESCRIPTION_32',
-      'HIRING_PAGE.JOB_DESCRIPTION_33',
-      'HIRING_PAGE.JOB_DESCRIPTION_34',
-      'HIRING_PAGE.JOB_DESCRIPTION_35',
-      'HIRING_PAGE.JOB_DESCRIPTION_10',
-      'HIRING_PAGE.JOB_DESCRIPTION_11',
-      'HIRING_PAGE.JOB_DESCRIPTION_12',
-      'HIRING_PAGE.JOB_DESCRIPTION_13',
-    ],
-    requirements: [
-      'HIRING_PAGE.JOB_REQUIREMENT_29',
-      'HIRING_PAGE.JOB_REQUIREMENT_30',
-      'HIRING_PAGE.JOB_REQUIREMENT_31',
-      'HIRING_PAGE.JOB_REQUIREMENT_32',
-      'HIRING_PAGE.JOB_REQUIREMENT_33',
-      'HIRING_PAGE.JOB_REQUIREMENT_34',
-      'HIRING_PAGE.JOB_REQUIREMENT_35',
-      'HIRING_PAGE.JOB_REQUIREMENT_36',
-      'HIRING_PAGE.JOB_REQUIREMENT_15',
-      'HIRING_PAGE.JOB_REQUIREMENT_4',
-      'HIRING_PAGE.JOB_REQUIREMENT_5',
-      'HIRING_PAGE.JOB_REQUIREMENT_7',
-    ],
-  },
-  {
-    anchor: 'job004',
-    jobTitle: 'HIRING_PAGE.JOB_NAME_4',
-    details: [
-      'HIRING_PAGE.JOB_DETAIL_PART_TIME',
-      'HIRING_PAGE.JOB_DETAIL_SALARY_3',
-      'HIRING_PAGE.JOB_DETAIL_LOCATION_1',
-    ],
-    descriptions: [
-      'HIRING_PAGE.JOB_DESCRIPTION_14',
-      'HIRING_PAGE.JOB_DESCRIPTION_15',
-      'HIRING_PAGE.JOB_DESCRIPTION_16',
-      'HIRING_PAGE.JOB_DESCRIPTION_17',
-      'HIRING_PAGE.JOB_DESCRIPTION_18',
-      'HIRING_PAGE.JOB_DESCRIPTION_19',
-    ],
-    requirements: [
-      'HIRING_PAGE.JOB_REQUIREMENT_18',
-      'HIRING_PAGE.JOB_REQUIREMENT_19',
-      'HIRING_PAGE.JOB_REQUIREMENT_20',
-      'HIRING_PAGE.JOB_REQUIREMENT_21',
-    ],
-  },
-  {
-    anchor: 'job005',
-    jobTitle: 'HIRING_PAGE.JOB_NAME_5',
-    details: [
-      'HIRING_PAGE.JOB_DETAIL_FULL_TIME',
-      'HIRING_PAGE.JOB_DETAIL_SALARY_4',
-      'HIRING_PAGE.JOB_DETAIL_LOCATION_1',
-    ],
-    descriptions: [
-      'HIRING_PAGE.JOB_DESCRIPTION_20',
-      'HIRING_PAGE.JOB_DESCRIPTION_21',
-      'HIRING_PAGE.JOB_DESCRIPTION_22',
-      'HIRING_PAGE.JOB_DESCRIPTION_23',
-      'HIRING_PAGE.JOB_DESCRIPTION_24',
-      'HIRING_PAGE.JOB_DESCRIPTION_25',
-      'HIRING_PAGE.JOB_DESCRIPTION_26',
-      'HIRING_PAGE.JOB_DESCRIPTION_17',
-      'HIRING_PAGE.JOB_DESCRIPTION_8',
-    ],
-    requirements: [
-      'HIRING_PAGE.JOB_REQUIREMENT_22',
-      'HIRING_PAGE.JOB_REQUIREMENT_23',
-      'HIRING_PAGE.JOB_REQUIREMENT_4',
-      'HIRING_PAGE.JOB_REQUIREMENT_24',
-      'HIRING_PAGE.JOB_REQUIREMENT_25',
-      'HIRING_PAGE.JOB_REQUIREMENT_26',
-      'HIRING_PAGE.JOB_REQUIREMENT_27',
-      'HIRING_PAGE.JOB_REQUIREMENT_28',
-    ],
-  },
-  {
-    anchor: 'job006',
-    jobTitle: 'HIRING_PAGE.JOB_NAME_6',
-    details: [
-      'HIRING_PAGE.JOB_DETAIL_FULL_TIME',
-      'HIRING_PAGE.JOB_DETAIL_SALARY_5',
-      'HIRING_PAGE.JOB_DETAIL_LOCATION_1',
-    ],
-    descriptions: [
-      'HIRING_PAGE.JOB_DESCRIPTION_36',
-      'HIRING_PAGE.JOB_DESCRIPTION_37',
-      'HIRING_PAGE.JOB_DESCRIPTION_38',
-      'HIRING_PAGE.JOB_DESCRIPTION_39',
-      'HIRING_PAGE.JOB_DESCRIPTION_40',
-      'HIRING_PAGE.JOB_DESCRIPTION_41',
-      'HIRING_PAGE.JOB_DESCRIPTION_42',
-      'HIRING_PAGE.JOB_DESCRIPTION_43',
-      'HIRING_PAGE.JOB_DESCRIPTION_44',
-      'HIRING_PAGE.JOB_DESCRIPTION_4',
-      'HIRING_PAGE.JOB_DESCRIPTION_5',
-      'HIRING_PAGE.JOB_DESCRIPTION_6',
-      'HIRING_PAGE.JOB_DESCRIPTION_7',
-      'HIRING_PAGE.JOB_DESCRIPTION_8',
-    ],
-    requirements: [
-      'HIRING_PAGE.JOB_REQUIREMENT_37',
-      'HIRING_PAGE.JOB_REQUIREMENT_19',
-      'HIRING_PAGE.JOB_REQUIREMENT_20',
-      'HIRING_PAGE.JOB_REQUIREMENT_21',
-      'HIRING_PAGE.JOB_REQUIREMENT_38',
-      'HIRING_PAGE.JOB_REQUIREMENT_39',
-      'HIRING_PAGE.JOB_REQUIREMENT_40',
-      'HIRING_PAGE.JOB_REQUIREMENT_41',
-      'HIRING_PAGE.JOB_REQUIREMENT_42',
-    ],
-  },
-  {
-    anchor: 'job007',
-    jobTitle: 'HIRING_PAGE.JOB_NAME_7',
-    details: [
-      'HIRING_PAGE.JOB_DETAIL_FULL_TIME',
-      'HIRING_PAGE.JOB_DETAIL_SALARY_5',
-      'HIRING_PAGE.JOB_DETAIL_LOCATION_1',
-    ],
-    descriptions: [
-      'HIRING_PAGE.JOB_DESCRIPTION_45',
-      'HIRING_PAGE.JOB_DESCRIPTION_46',
-      'HIRING_PAGE.JOB_DESCRIPTION_47',
-      'HIRING_PAGE.JOB_DESCRIPTION_48',
-      'HIRING_PAGE.JOB_DESCRIPTION_49',
-      'HIRING_PAGE.JOB_DESCRIPTION_41',
-      'HIRING_PAGE.JOB_DESCRIPTION_42',
-      'HIRING_PAGE.JOB_DESCRIPTION_50',
-      'HIRING_PAGE.JOB_DESCRIPTION_51',
-      'HIRING_PAGE.JOB_DESCRIPTION_4',
-      'HIRING_PAGE.JOB_DESCRIPTION_5',
-      'HIRING_PAGE.JOB_DESCRIPTION_6',
-      'HIRING_PAGE.JOB_DESCRIPTION_7',
-      'HIRING_PAGE.JOB_DESCRIPTION_8',
-    ],
-    requirements: [
-      'HIRING_PAGE.JOB_REQUIREMENT_43',
-      'HIRING_PAGE.JOB_REQUIREMENT_19',
-      'HIRING_PAGE.JOB_REQUIREMENT_20',
-      'HIRING_PAGE.JOB_REQUIREMENT_21',
-      'HIRING_PAGE.JOB_REQUIREMENT_44',
-      'HIRING_PAGE.JOB_REQUIREMENT_45',
-      'HIRING_PAGE.JOB_REQUIREMENT_46',
-      'HIRING_PAGE.JOB_REQUIREMENT_47',
-      'HIRING_PAGE.JOB_REQUIREMENT_40',
-      'HIRING_PAGE.JOB_REQUIREMENT_41',
-      'HIRING_PAGE.JOB_REQUIREMENT_42',
-    ],
-  },
-];
-
 export const mermerAddressInMap =
   'https://www.google.com/maps/place/105%E5%8F%B0%E5%8C%97%E5%B8%82%E6%9D%BE%E5%B1%B1%E5%8D%80%E6%B0%91%E7%94%9F%E6%9D%B1%E8%B7%AF%E4%B8%89%E6%AE%B5156%E8%99%9F/@25.0573596,121.5455644,17z/data=!4m16!1m9!3m8!1s0x3442abe5fa651b97:0x102228e5b00fa9d3!2zMTA15Y-w5YyX5biC5p2-5bGx5Y2A5rCR55Sf5p2x6Lev5LiJ5q61MTU26Jmf!3b1!8m2!3d25.0573596!4d121.5477531!10e5!16s%2Fg%2F11j5999y35!3m5!1s0x3442abe5fa651b97:0x102228e5b00fa9d3!8m2!3d25.0573596!4d121.5477531!16s%2Fg%2F11j5999y35?entry=ttu';
